backend/apps/auth/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic == 2.*
PyJWT==2.8.0